add: hero section
